pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "engram"
version = "0.1.0"
description = "Framework-agnostic agent memory layer with policies, observability, and pluggable storage"
requires-python = ">=3.9"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools]
packages = ["engram"]
